+++ scanpy/tests/external/test_harmony_integrate.py
@@ -1,0 +1,21 @@
+import scanpy as sc
+import scanpy.external as sce
+from scanpy.testing._helpers.data import pbmc3k
+from scanpy.testing._pytest.marks import needs
+
+
+@needs("harmonypy")
+def test_harmony_integrate():
+    """
+    Test that Harmony integrate works.
+
+    This is a very simple test that just checks to see if the Harmony
+    integrate wrapper succesfully added a new field to ``adata.obsm``
+    and makes sure it has the same dimensions as the original PCA table.
+    """
+    adata = pbmc3k()
+    sc.pp.recipe_zheng17(adata)
+    sc.tl.pca(adata)
+    adata.obs['batch'] = 1350 * ['a'] + 1350 * ['b']
+    sce.pp.harmony_integrate(adata, 'batch')
+    assert adata.obsm['X_pca_harmony'].shape == adata.obsm['X_pca'].shape
